+++ dialect_common.go
@@ -199,7 +199,7 @@ func IsByteArrayOrSlice(value reflect.Value) bool {
 }
 
 // Returns the primary key via the row ID
-func (commonDialect) ResolveRowID(tableName string, rowID uint) uint{
+func (commonDialect) ResolveRowID(tableName string, rowID uint64) uint64{
 	// In case of most DB's, it's assumed that the result of the `result.LastInsertId()` is the ID of the record
 	return rowID
 }
